+++ README.md
@@ -44,6 +44,10 @@ PDF.js is built into version 19+ of Firefox.
 Chrome, go to `Tools > Extension` and load the (unpackaged) extension from the
 directory `build/chromium`.
 
+### PDF debugger
+
+Browser the internal structure of a PDF document with https://mozilla.github.io/pdf.js/internal-viewer/web/pdf_internal_viewer.html
+
 ## Getting the Code
 
 To get a local copy of the current code, clone it using git:
